// src/app/heroes/interfaces/hero.interface.ts
// Generated by https://quicktype.io

export interface Hero {
    id:               string;
    superhero:        string;
    publisher:        Publisher;
    alter_ego:        string;
    first_appearance: string;
    characters:       string;
    alt_img?:         string;
}

//Enums are for only specific types of values? "This or That or...." etc
export enum Publisher {
    DCComics = "DC Comics",
    MarvelComics = "Marvel Comics",
}